Created file structure as per comment provided, added tc-id , resolved code-rabbit review
