+++ tests/workspace/create-workspace/create-workspace.spec.ts
@@ -25,7 +25,7 @@ function findCreatedWorkspaceDirs(location: string): string[] {
 
 test.describe('Create Workspace', () => {
   test.describe('Inline Creation Flow', () => {
-    test('should create workspace via inline rename and press Enter', { tag: '@sanity' }, async ({ launchElectronApp, createTmpDir }) => {
+    test('TC-957: Verify user able to Create a Workspace', { tag: '@sanity' }, async ({ launchElectronApp, createTmpDir }) => {
       const wsLocation = await createTmpDir('ws-location-enter');
 
       const app = await launchElectronApp({ initUserDataPath, templateVars: { wsLocation } });
@@ -454,7 +454,7 @@ test.describe('Create Workspace', () => {
       await closeElectronApp(app);
     });
 
-    test('should persist workspace name after app restart', { tag: '@sanity' }, async ({ launchElectronApp, createTmpDir }) => {
+    test('TC-959: Verify created Workspace persists even after bruno app restart.', { tag: '@sanity' }, async ({ launchElectronApp, createTmpDir }) => {
       const userDataPath = await createTmpDir('create-ws-name-persist');
       const wsLocation = await createTmpDir('ws-location-persist');
